Update file_2_preprocessing.py
extra comment to rotation matrix function. 

Please double check this function. There might be a problem with the calculation. The angle is computed using absolute differences (dx and dy). This might lead to an issue because the arctan2 function typically uses the actual differences (which can be negative) to determine the quadrant. However, in this code, dx and dy are absolute values. This is a critical point because using absolute values will always place the point in the first quadrant (both dx and dy positive). Then the angle will always be between 0 and 90 degrees.

If we are trying to align the vector from origin to the_other to the x-axis, then we should compute the angle without absolute values:
   dx = float(the_other['X']) - float(origin['X'])
   dy = float(the_other['Y']) - float(origin['Y']) 

I am not 100% sure that my thoughts are correct. So please double chekc the code and my try with some random values